+++ include/parser.h
@@ -3,43 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   parser.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emonacho <emonacho@student.42lausanne.c    +#+  +:+       +#+        */
+/*   By: timmi <timmi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/18 17:50:04 by emonacho          #+#    #+#             */
-/*   Updated: 2025/04/20 22:07:45 by emonacho         ###   ########.fr       */
+/*   Updated: 2025/04/24 10:43:59 by timmi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef PARSER_H
 # define PARSER_H
-
-// ⚠️👷‍♂️ Travail en cours, fonctions et structures probablement à modifier
-
-// Adapté de: https://keleshev.com/abstract-syntax-tree-an-example-in-c/
-typedef enum	e_tag
-{
-	AST_BLOCK,
-	AST_PIPE,
-	AST_LINE,
-	AST_REDIR,
-	AST_EXEC
-} 				t_tag;
-
-// FORWARD DECLARATION (dis au compilateur que "t_ast" existe)
-typedef struct s_ast t_ast;
-
-typedef struct	s_ast
-{
-	t_tag		tag;
-	union		u_data
-	{
-		struct { t_ast *left; t_ast *right; } ast_block;
-		struct { t_ast *left; t_ast *right; } ast_pipe;
-		struct { t_ast *left; t_ast *right; } ast_line;
-		struct { t_ast *left; t_ast *right; } ast_redir;
-		struct { int argc; char **argv; } ast_exec;
-	}			data;
-}				t_ast;
 
 // parser.c
 t_ast	*parse_tokens(t_list **head);
