+++ lib/libft/ft_strnprefix.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strnprefix.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: c4v3d <c4v3d@student.42.fr>                +#+  +:+       +#+        */
+/*   By: timmi <timmi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/22 11:11:30 by c4v3d             #+#    #+#             */
-/*   Updated: 2025/04/22 11:13:32 by c4v3d            ###   ########.fr       */
+/*   Updated: 2025/04/24 09:58:16 by timmi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
