+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: c4v3d <c4v3d@student.42.fr>                +#+  +:+       +#+        */
+/*   By: timmi <timmi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 09:49:18 by timmi             #+#    #+#             */
-/*   Updated: 2025/04/23 09:03:50 by c4v3d            ###   ########.fr       */
+/*   Updated: 2025/04/24 10:38:50 by timmi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,28 +14,30 @@
 
 void initialize_struct(t_shell *s)
 {
-	(void)s;
+	s->env = NULL;
+	s->cmd_count = 0;
+	s->line = NULL;
+	s->head = NULL;
+	s->root_node = NULL;
 }
 
 void prompt_loop(char *prompt, t_shell *s)
 {
-	char *line_read;
 	int loop;
 
 	loop = 1;
 	while (loop)
 	{
-		line_read = readline(prompt);
-		if (line_read && *line_read) // need to add a check to not print strings containing only spaces
+		s->line = readline(prompt);
+		if (s->line) // need to add a check to not print strings containing only spaces
 		{
-			add_history(line_read);
-			s->head = tokenize(line_read);
-			exit_check(s);
+			add_history(s->line);
+			lexer(s);
 			s->root_node = parse_tokens(&s->head);
 			free_list(s->head);
 		}
 		// (void)ast; // 💥TEST
-		free(line_read);
+		free(s->line);
 	}
 }
 
@@ -43,10 +45,7 @@ int main(int argc, char **argv)
 {
 	t_shell s;
 	char *prompt;
-
-	s.env = NULL;
-	s.head = NULL;
-	s.root_node = NULL;
+	
 	if (argc > 1)
 	{
 		printf("\nEntering Debug mode !\n\n");
@@ -55,6 +54,7 @@ int main(int argc, char **argv)
 	else
 	{
 		prompt = create_prompt();
+		initialize_struct(&s);
 		prompt_loop(prompt, &s);
 		free(prompt);
 	}
